WEB-3920: move schedule flow components to tasks folder
