+++ app/(candidate)/dashboard/tasks/components/flows/CloseConfirmModal.js
