+++ app/(candidate)/dashboard/tasks/components/flows/ScheduleAddScriptFlow/ScheduleAddScriptFlow.js
@@ -1,12 +1,12 @@
 'use client'
 import { useState } from 'react'
-import { ChooseScriptAddFlow } from 'app/(candidate)/dashboard/voter-records/[type]/components/ScheduleAddScriptFlow/ChooseScriptAddFlow'
-import { ADD_SCRIPT_FLOW } from 'app/(candidate)/dashboard/voter-records/[type]/components/ScheduleAddScriptFlow/AddScriptFlow'
-import { SelectSmSScriptScreen } from 'app/(candidate)/dashboard/voter-records/[type]/components/ScheduleAddScriptFlow/SelectSmSScriptScreen'
-import { CreateSmSScriptScreen } from 'app/(candidate)/dashboard/voter-records/[type]/components/ScheduleAddScriptFlow/CreateSmSScriptScreen'
-import { SelectSmsAiTemplateScreen } from 'app/(candidate)/dashboard/voter-records/[type]/components/ScheduleAddScriptFlow/SelectSmsAiTemplateScreen'
-import { GenerateLoadingScreen } from 'app/(candidate)/dashboard/voter-records/[type]/components/ScheduleAddScriptFlow/GenerateLoadingScreen'
-import { GenerateReviewScreen } from 'app/(candidate)/dashboard/voter-records/[type]/components/ScheduleAddScriptFlow/GenerateReviewScreen'
+import { ChooseScriptAddFlow } from './ChooseScriptAddFlow'
+import { ADD_SCRIPT_FLOW } from './AddScriptFlow'
+import { SelectSmSScriptScreen } from './SelectSmSScriptScreen'
+import { CreateSmSScriptScreen } from './CreateSmSScriptScreen'
+import { SelectSmsAiTemplateScreen } from './SelectSmsAiTemplateScreen'
+import { GenerateLoadingScreen } from './GenerateLoadingScreen'
+import { GenerateReviewScreen } from './GenerateReviewScreen'
 
 export default function ScheduleAddScriptFlow({
   onComplete = (scriptKey) => {},
