+++ app/(candidate)/dashboard/tasks/components/flows/ScheduleAddScriptFlow/SelectSmSScriptScreen.js
@@ -1,7 +1,7 @@
 import { useState } from 'react'
 import H1 from '@shared/typography/H1'
 import Body1 from '@shared/typography/Body1'
-import { SmsScriptSelect } from 'app/(candidate)/dashboard/voter-records/[type]/components/ScheduleAddScriptFlow/SmsScriptSelect'
+import { SmsScriptSelect } from './SmsScriptSelect'
 import { ModalFooter } from '@shared/ModalFooter'
 import { trackEvent, EVENTS } from 'helpers/fullStoryHelper'
