+++ app/(candidate)/dashboard/tasks/components/flows/ScheduleAddScriptFlow/SelectSmsAiTemplateScreen.js
@@ -5,7 +5,7 @@ import Body1 from '@shared/typography/Body1'
 import { ModalFooter } from '@shared/ModalFooter'
 import { useEffect, useState } from 'react'
 import Link from 'next/link'
-import { SmsTemplateSelect } from 'app/(candidate)/dashboard/voter-records/[type]/components/ScheduleAddScriptFlow/SmsTemplateSelect'
+import { SmsTemplateSelect } from './SmsTemplateSelect'
 import { setRequiresQuestionsOnTemplates } from 'helpers/setRequiresQuestionsOnTemplates'
 import { clientFetch } from 'gpApi/clientFetch'
 import { apiRoutes } from 'gpApi/routes'
